refactor(019/extract-scan): move single-file binary scanner from binary/mod.rs to scan.rs — completes the split
Phase 5 of milestone 019 — extracts the single-file scanner cohort
(is_go_binary, scan_binary, scan_fat_macho, collect_string_region) plus
the four is_go_binary regression tests into a new sibling module.

scan.rs imports BinaryScan from sibling entry.rs (cross-sibling type
import), per the visibility ladder + research R3.

mod.rs final state: 449 LOC (down from 1858 baseline), well under the
FR-001 ≤800 ceiling. Goldens regen clean (zero diff in 27 byte-identity
goldens). External callers (scan_fs/mod.rs, package_db/maven.rs +
go_binary.rs, binary/linkage.rs) untouched (SC-005 verified empty diff).

Per-submodule LOC inventory:
- binary/mod.rs:        449 (orchestration + is_path_claimed + integration tests)
- binary/predicates.rs: 303 (rootfs detection + OS predicates + 14 tests)
- binary/discover.rs:    89 (filesystem walker + format detection)
- binary/entry.rs:      579 (BinaryScan struct + entry conversions + 13 tests)
- binary/scan.rs:       306 (scan_binary + scan_fat_macho + 4 tests)
